chore: Update client and user DTOs and entities

+++ src/client/client.service.ts
@@ -1,11 +1,12 @@
-// src/client/client.service.ts
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
+import { Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Client } from './entities/client.entity';
-import { CreateClientDto } from './dto/create-client.dto';
+import { RegisterClientDto } from './dto/register-client.dto';
 import { AwsService } from '../aws/aws.service';
+import { UserService } from '../user/user.service';
 import * as bcrypt from 'bcrypt';
+import { CreateUserDto } from './../auth/dto/create-user.dto';
 
 @Injectable()
 export class ClientService {
@@ -13,37 +14,68 @@ export class ClientService {
     @InjectRepository(Client)
     private clientsRepository: Repository<Client>,
     private readonly awsService: AwsService,
+    private readonly userService: UserService,
   ) {}
 
-  async create(createClientDto: CreateClientDto): Promise<Client> {
-    const client = this.clientsRepository.create(createClientDto);
+  async create(registerClientDto: RegisterClientDto): Promise<Client> {
+    const hashedPassword = await bcrypt.hash(registerClientDto.password, 10);
 
-    // Hash the password
-    client.password = await bcrypt.hash(createClientDto.password, 10);
+    // Create a User entity for the client
+    const userDto: CreateUserDto = {
+      username: registerClientDto.name,
+      email: registerClientDto.email,
+      password: registerClientDto.password,
+      password_hash: hashedPassword,
+      role: 'client',
+      is_client_admin: true,
+    };
+    const user = await this.userService.create(userDto);
 
-    await this.clientsRepository.save(client);
-    await this.awsService.createS3Folder(client.id);
-    return client;
+    const client = this.clientsRepository.create({
+      ...registerClientDto,
+      user: user,
+    });
+
+    try {
+      await this.clientsRepository.save(client);
+      await this.awsService.createS3Folder(client.id);
+      return client;
+    } catch (error) {
+      throw new InternalServerErrorException('Error creating client');
+    }
   }
 
   async findAll(): Promise<Client[]> {
-    return this.clientsRepository.find();
+    return this.clientsRepository.find({ relations: ['user'] });
   }
 
   async findOne(id: string): Promise<Client> {
-    return this.clientsRepository.findOne({ where: { id } });
+    const client = await this.clientsRepository.findOne({ where: { id }, relations: ['user'] });
+    if (!client) {
+      throw new NotFoundException('Client not found');
+    }
+    return client;
   }
 
   async findOneByEmail(email: string): Promise<Client | undefined> {
-    return this.clientsRepository.findOne({ where: { email } });
+    return this.clientsRepository.findOne({ where: { email }, relations: ['user'] });
   }
 
   async update(id: string, updateClientDto: Partial<Client>): Promise<Client> {
-    await this.clientsRepository.update(id, updateClientDto);
-    return this.findOne(id);
+    const client = await this.findOne(id);
+    if (!client) {
+      throw new NotFoundException('Client not found');
+    }
+    Object.assign(client, updateClientDto);
+    await this.clientsRepository.save(client);
+    return client;
   }
 
   async remove(id: string): Promise<void> {
+    const client = await this.findOne(id);
+    if (!client) {
+      throw new NotFoundException('Client not found');
+    }
     await this.clientsRepository.delete(id);
   }
 }
